refactor(frontend): rename network 'Routes' files
